apps/dspy/requirements.txt
# Common packages for all our azurefed apps
python-dotenv
streamlit
semantic-kernel
msal

# App sepcific packages
